BigInt for users

+++ app/models.py
@@ -6,7 +6,7 @@ class Size(Base):
     __tablename__ = 'sizes'
 
     chat_id = Column(BigInteger, primary_key=True)
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     size = Column(Integer, default=0)
     last_update = Column(DateTime, onupdate=func.now(), server_default=func.now())
     isUpdated = Column(Boolean, default=False)
